InputEncoder

+++ engine/src/scribblez/agent.cpp
@@ -19,7 +19,7 @@ GreedyAgent::GreedyAgent(int thread_id, const std::string& name, uint64_t seed)
 Move GreedyAgent::make_move(const MoveRequest& req) {
   if (!req.legal_plays.empty()) {
     int best = req.legal_plays.front().score;
-    for (const auto& m : req.legal_plays) best = std::max(best, m.score);
+    for (const auto& m : req.legal_plays) best = std::max(best, static_cast<int>(m.score));
     std::vector<const Move*> top;
     for (const auto& m : req.legal_plays)
       if (m.score == best) top.push_back(&m);
